refactor(registry): extract state helpers into cartog-registry
`slot_for_db` and `default_state_dir` lived in the `cartog` binary crate,
where `cartog-mcp` and `cartog-watch` could not reach them — the binary
depends on those crates, not the reverse. Both documented the binary's
function as the way to derive the PID slot they lock on while being unable
to call it. Move them into a new `cartog-registry` crate (11th published)
so those references become real.

Move-only. `slot_for_db`, `resolve_db_path_for_slot`, `detect_live_serve_peer`,
`default_state_dir` and `default_state_file` move verbatim (one const rename,
FILE_NAME -> STATE_FILE_NAME, since the module is no longer state.toml-only),
along with all 17 of their tests. `State`/`PendingUpdate`/`load_from`/`save_to`
stay behind as the serde+toml island. `state.rs` re-exports the moved names so
the in-crate call sites and `cartog::state::*` keep resolving unchanged.

Six seam tests cover the invariant that now spans a crate boundary: the two
crates' `slot_for_db` are one function, equivalent and symlinked paths still
collapse onto one slot, a generated slot round-trips through the PID-lock
writer, and `detect_live_serve_peer` still matches a generated serve slot.

Also repays the doc debt the move exists to fix: the 8 dangling
`cartog::state::*` references in cartog-watch and cartog-mcp now name
`cartog_registry::slot_for_db`.

`release.yml` gains the crate in publish order (after cartog-process-lock,
its only internal dep, before cartog-watch) — omitting it would break the
next release with no local signal.

The two design-proposal docs on this branch are committed as-is; the
registry store they propose is not implemented here.
