+++ docs/README.md
@@ -31,6 +31,8 @@ This documentation is organized by the [Diataxis](https://diataxis.fr) framework
 - [explanation/rag-pipeline.md](explanation/rag-pipeline.md) — hybrid search, embedding models, provider architecture
 - [explanation/incremental-indexing.md](explanation/incremental-indexing.md) — three-layer change detection, Merkle invariants
 - [explanation/concurrency.md](explanation/concurrency.md) — rayon / tokio / std::thread models
+- [explanation/project-registry.md](explanation/project-registry.md) — *proposal* — user-global project registry and cross-project discovery
+- [explanation/cross-project-queries.md](explanation/cross-project-queries.md) — *proposal* — federated search and cross-service edges
 
 ## Background docs
